+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.10.1 (Next)
 
+* [#227](https://github.com/tim-vandecasteele/grape-swagger/issues/227): Fix: nested routes under prefix not documented - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.10.0 (March 10, 2015)